requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=24.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
